requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.22.2
python-decouple==3.3
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
